services/biometric/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
deepface==0.0.89
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
tf-keras==2.16.0
